# src/test/resources/api/openapi.yaml
openapi: 3.0.1
info:
  title: api-oep-integrator
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
  /{municipalityId}/{instanceType}/cases/{flowInstanceId}/status:
    get:
      tags:
      - Cases
      summary: Get case status
      description: Get case status by flow instance ID
      operationId: getCaseStatusByFlowInstanceId
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: instanceType
        in: path
        description: The instanceType where case belongs
        required: true
        schema:
          $ref: "#/components/schemas/InstanceType"
        example: INTERNAL
      - name: flowInstanceId
        in: path
        description: flow-instance id
        required: true
        schema:
          type: string
        example: 112233
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStatus"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
    put:
      tags:
      - Cases
      summary: Set case status
      description: Set case status by flowInstanceId
      operationId: setStatus
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: instanceType
        in: path
        description: The instanceType where case belongs
        required: true
        schema:
          $ref: "#/components/schemas/InstanceType"
        example: INTERNAL
      - name: flowInstanceId
        in: path
        description: flow-instance ID
        required: true
        schema:
          type: string
        example: 112233
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseStatusChangeRequest"
        required: true
      responses:
        "204":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStatusChangeResponse"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/{instanceType}/cases/systems/{system}/{externalId}/status:
    put:
      tags:
      - Cases
      summary: Set case status
      description: Set case status by externalId
      operationId: setStatus_1
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: instanceType
        in: path
        description: The instanceType where case belongs
        required: true
        schema:
          $ref: "#/components/schemas/InstanceType"
        example: INTERNAL
      - name: system
        in: path
        description: The system where external ID exists
        required: true
        schema:
          type: string
        example: CaseData
      - name: externalId
        in: path
        description: Case ID in specified system
        required: true
        schema:
          type: string
        example: 234
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CaseStatusChangeRequest"
        required: true
      responses:
        "204":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStatusChangeResponse"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/{instanceType}/cases/{flowInstanceId}/queries/{queryId}/files/{fileId}:
    get:
      tags:
      - Cases
      summary: Get case attachment
      description: "Get case attachment by flowInstanceId, queryId and fileId"
      operationId: getCaseAttachment
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: instanceType
        in: path
        description: The instanceType where case belongs
        required: true
        schema:
          $ref: "#/components/schemas/InstanceType"
        example: INTERNAL
      - name: flowInstanceId
        in: path
        description: flow-instance ID
        required: true
        schema:
          type: string
        example: 112233
      - name: queryId
        in: path
        description: Query ID
        required: true
        schema:
          type: string
        example: 112233
      - name: fileId
        in: path
        description: File ID
        required: true
        schema:
          type: string
        example: 112233
      responses:
        "200":
          description: Successful operation
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/{instanceType}/cases/{flowInstanceId}/pdf:
    get:
      tags:
      - Cases
      summary: Get case PDF
      description: Get case PDF by flow instance ID
      operationId: getCasePdfByFlowInstanceId
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: instanceType
        in: path
        description: The instanceType where case belongs
        required: true
        schema:
          $ref: "#/components/schemas/InstanceType"
        example: INTERNAL
      - name: flowInstanceId
        in: path
        description: flow-instance ID
        required: true
        schema:
          type: string
        example: 112233
      responses:
        "200":
          description: Successful operation
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/{instanceType}/cases/parties/{partyId}:
    get:
      tags:
      - Cases
      summary: Get cases by citizen identifier
      description: Get a list of case envelopes by citizen identifier
      operationId: getCasesByPartyId
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: instanceType
        in: path
        description: The instanceType where case belongs
        required: true
        schema:
          $ref: "#/components/schemas/InstanceType"
        example: INTERNAL
      - name: partyId
        in: path
        description: The party ID
        required: true
        schema:
          type: string
        example: 123
      - name: fromDate
        in: query
        description: Filter cases on fromDate
        required: false
        schema:
          type: string
          format: date
        example: 2024-01-01
      - name: toDate
        in: query
        description: Filter cases on toDate
        required: false
        schema:
          type: string
          format: date
        example: 2024-01-31
      - name: status
        in: query
        description: Filter by status
        required: false
        schema:
          type: string
        example: Preliminär
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaseEnvelope"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/{instanceType}/cases/families/{familyId}:
    get:
      tags:
      - Cases
      summary: Get cases by family ID
      description: Get a list of case envelopes by family ID
      operationId: getCasesByFamilyId
      parameters:
      - name: municipalityId
        in: path
        description: Municipality ID
        required: true
        schema:
          type: string
        example: 2281
      - name: instanceType
        in: path
        description: The instanceType where case belongs
        required: true
        schema:
          $ref: "#/components/schemas/InstanceType"
        example: INTERNAL
      - name: familyId
        in: path
        description: The family ID
        required: true
        schema:
          type: string
        example: 123
      - name: fromDate
        in: query
        description: Filter cases on fromDate
        required: false
        schema:
          type: string
          format: date
        example: 2024-01-01
      - name: toDate
        in: query
        description: Filter cases on toDate
        required: false
        schema:
          type: string
          format: date
        example: 2024-01-31
      - name: status
        in: query
        description: Filter by status
        required: false
        schema:
          type: string
        example: Preliminär
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaseEnvelope"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "502":
          description: Bad gateway
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        detail:
          type: string
        title:
          type: string
    StatusType:
      type: object
      properties:
        reasonPhrase:
          type: string
        statusCode:
          type: integer
          format: int32
    InstanceType:
      type: string
      description: InstanceType type. INTERNAL/EXTERNAL
      example: INTERNAL
      enum:
      - INTERNAL
      - EXTERNAL
    CaseStatusChangeRequest:
      type: object
      properties:
        id:
          type: integer
          description: The status ID
          format: int32
          example: 123
        name:
          type: string
          description: The status name
          example: Inskickat
        principal:
          $ref: "#/components/schemas/Principal"
      description: SetStatus request model
    Principal:
      required:
      - userId
      type: object
      properties:
        name:
          type: string
          description: The name
          example: John Doe
        userId:
          type: string
          description: The user id
          example: joh12doe
      description: Principal model
    ConstraintViolationProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        type:
          type: string
          format: uri
        status:
          $ref: "#/components/schemas/StatusType"
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        message:
          type: string
        instance:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    ThrowableProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        message:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        detail:
          type: string
        title:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
    CaseStatusChangeResponse:
      type: object
      properties:
        eventId:
          type: integer
          description: Event id
          format: int32
          example: 123
      description: SetStatus response model
...
    CaseStatus:
      type: object
      properties:
        id:
          type: integer
          description: The status ID
          format: int32
          example: 123
        name:
          type: string
          description: The status name
          example: Inskickat
      description: Case status model
    CaseEnvelope:
      type: object
      properties:
        flowInstanceId:
          type: string
          description: The flowInstanceId (caseId)
          example: "109581"
        created:
          type: string
          description: The case create date
          format: date-time
        statusUpdated:
          type: string
          description: The case status change date
          format: date-time
      description: Case envelope model
  securitySchemes: {}